now with queues, i hope

+++ common/utility/Constants.h
@@ -17,9 +17,9 @@
 //ethertype for ethernet header used in wifi commuication
 static const uint16_t ETHERTYPE_CAR = 0x0CA4;
 
-static const uint16_t PRIORITY_VO =6;
-static const uint16_t PRIORITY_VI =4;
-static const uint16_t PRIORITY_BE =0;
-static const uint16_t PRIORITY_BK =1;
+static const int PRIORITY_VO =6;
+static const int PRIORITY_VI =4;
+static const int PRIORITY_BE =0;
+static const int PRIORITY_BK =1;
 
 #endif /* UTILITY_CONSTANTS_H_ */
